Issue #102 Implement user hot-reloading
The config file's directory has to be watched, too, because some editors
(vim) replace the file in a way that tiny-auth otherwise loses track of
it.

Allow fallible parsing of the configuration file without exiting the
process inside the parsing.

Broadcast changed files to all stores. The store determines by itself if
it is responsible for a file.