Issue #5 Template SQL queries
Allow data loaders to use all loaded data as well as the special
variable tiny_auth_assigned_to as tera templates in SQL queries.

Loaded data is available as object if the chain of destinations of data
loaders consists only of "to one" multiplicities. It is available as
array otherwise.