Issue #105 Clean test architecture
In order to not have test fixtures in productive code, extract these
modules into a separate crate, only used in dev-dependencies.

This causes a problem for business, because it creates duplicate,
incompatible types because business is compiled twice.
Avoid this by promoting business tests using test fixtures to
integration tests. Those cannot use the private api of the crate, but
this is no issue atm.

Don't migrate dev db for test setup for speed.